[RPC] Ensure that a numeric is being passed to AmmountFromValue
Return a more descriptive type error when the input paramater is not a
numeric.

This clarifies the response message a bit and allows for a faster exit
response by not needing to do subsequent mathematical or external
function checks.

NOTE: This isn't a functional change, since non-numeric input values
would previously fail anyways, returning only an "Invalid amount" error
message.

https://github.com/PIVX-Project/PIVX/pull/724/commits/1f56fbaf472f50e1a96da4c93f1723c00a299ad8