DAPS-1305 from PIVX: Improve addrman Select() performance when buckets are nearly empty
Simple backport of bitcoin#6530

PIVX-Project/PIVX#798

Earlier, when there are only a few entries in the addrman tables, Select() could spin for long periods of time (~200ms), resulting in high CPU usage (up to 40% here). This reduces it to around 0.3ms (0.5% CPU).

(We have half of this commit for some reason, second half added)