DAPS-1330 - Multisig [commit by commit] (v1.0.4.4-v1.0.4.6) (#973)
* DAPS-1206 - Force users to re-enter their password to reveal their mnemonic phrase

* DAPS-761 DAPS-1296 From PIVX: Unify shutdown procedure in init rather than per-app

This moves CScheduler and threadGroup to a static declaration in init
.cpp so as to avoid a potential shutdown deadlock where both are freed
before the network message handler thread has been completely released.

* DAPS-1193 Move unlock-dialog from splash screen to overview screen

* Bump version to v1.0.4.5

* reenable invalidateblock and reconsiderblock RPC

* DAPS-1248 from Bitcoin: net: initialize socket to avoid closing random fd's

bitcoin/bitcoin#12326
"An excellent spot by @david60.

Even if it isn't causing the fd issue we're looking for, this should be fixed."

* DAPS-761 From PIVX Upstream: Db runtime error cleaning the variable that needs to be logged

"Straightforward change, the variable is been cleaned before the logging. Temp variable copy to be able to print it."

Could be useful for future logging

from: PIVX-Project/PIVX#989

* DAPS-1334: Fix Runaway Exception - Move the RandAddSeedPerfmon() call into the MakeNewKey function

"CKey::MakeNewKey is responsible for generating a new private key using a cryptographic PRNG. The rest of key metadata is then generated by CWallet::GenerateNewKey.

However, the GenerateNewKey function does not call MakeNewKey before ensuring RandAddSeedPerfmon is done. If increasing the uncertainty about the state and making the PRNG output less predictable is the message to send, the step then might as well be implemented in the MakeNewKey function itself.

The initial Sanity Check, and tests are the two other consumers of the function, neither RandAddSeedPerfmon-ing. The latter out of, presumably, performance considerations. Nevertheless, the CKey access modifiers, and the ongoing libification tipped me over to log this, as it is something to consider as we move forward (in light of recent events...)."

bitcoin/bitcoin#5508
bitcoin/bitcoin#5495

* DAPS-1342 From PIVX: WIN32 Seed Cleanup: Move nLastPerfmon behind win32 ifdef.

"Code to avoid calling Perfmon too often is only needed when perfmon is actually going to get called.
This is not intended to make any functional difference in the addition of entropy to the random pool."

From PIVX-Project/PIVX#762

* DAPS-1340 From PIVX: Remove Bitcoin Core 0.8 block hardlinking

Will never be used in our case

from: PIVX-Project/PIVX#980

* use block time for transaction time

* DAPS-1333 - Windows Installer: Remove testnet shortcut

This is to eliminate some confusion for the users using the Windows Installer - a shortcut to launch Testnet was being created and added to the start menu.

* DAPS-1318 Fix for memory deallocation bug during ungraceful exit

* DAPS-1112 Dont auto-ban whitelisted nodes

* DAPS-1328 from Bitcoin: net: don't retry failed oneshot connections forever

"As introduced by (my suggestion, sorry, in) #11512, failed dns resolves end up as oneshots. But failed oneshots are re-added as oneshots, so we need to make sure that we're not queuing these up forever after failed resolves.

Rather than trying to differentiate, I think we should just not re-add failed oneshots and be done with it.

Maybe @sipa can shed a light on what the original intention was."

* Bump version to v1.0.4.6